Refactored caching.
