+++ src/skshapes/data/__init__.py
@@ -8,6 +8,7 @@ import torch
 
 from ..input_validation import typecheck
 from ..types import polydata_type
+from .data_attributes import DataAttributes
 from .image import Image
 from .polydata import PolyData
